+++ Realization/frontend/czechidm-core/src/components/advanced/ContractPositionInfo/ContractPositionInfo.js
@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 //
 import * as Utils from '../../../utils';
@@ -16,10 +17,6 @@ const manager = new ContractPositionManager();
  * @author Radek Tomiška
  */
 export class ContractPositionInfo extends AbstractEntityInfo {
-
-  constructor(props, context) {
-    super(props, context);
-  }
 
   getManager() {
     return manager;